Move async-store back under src/area-chart
