+++ src/area-chart/elements/use-highlight-details.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { useSelector } from '../../internal/async-store';
+import { useSelector } from '../async-store';
 import { CartesianChartProps } from '../../internal/components/cartesian-chart/interfaces';
 import { ChartSeriesDetailItem } from '../../internal/components/chart-series-details';
 import { useInternalI18n } from '../../internal/i18n/context';
